Update tests that show the feature is experimentally gated
